Adicionado o view e deixando o main limpo
